--- backend/src/main/java/com/paichai/health/pt_session/controller/PtSessionController.java
@@ -1,5 +1,0 @@
-package com.paichai.health.pt_session.controller;
-
-public class PtSessionController {
-
-}
