--- backend/src/main/java/com/paichai/health/pt_session/dto/PtSessionRequest.java
@@ -1,5 +1,0 @@
-package com.paichai.health.pt_session.dto;
-
-public class PtSessionRequest {
-
-}
