--- backend/src/main/java/com/paichai/health/pt_session/entity/PtSession.java
@@ -1,5 +1,0 @@
-package com.paichai.health.pt_session.entity;
-
-public class PtSession {
-
-}
