--- backend/src/main/java/com/paichai/health/pt_session/repository/PtSessionRepository.java
@@ -1,5 +1,0 @@
-package com.paichai.health.pt_session.repository;
-
-public class PtSessionRepository {
-
-}
