--- backend/src/main/java/com/paichai/health/pt_session/service/PtSessionService.java
@@ -1,5 +1,0 @@
-package com.paichai.health.pt_session.service;
-
-public class PtSessionService {
-
-}
