+++ backend/src/main/java/com/paichai/health/ptsession/controller/PtSessionController.java
@@ -1,0 +1,5 @@
+package com.paichai.health.ptsession.controller;
+
+public class PtSessionController {
+
+}
