+++ backend/src/main/java/com/paichai/health/ptsession/dto/PtSessionRequest.java
@@ -1,0 +1,5 @@
+package com.paichai.health.ptsession.dto;
+
+public class PtSessionRequest {
+
+}
