+++ backend/src/main/java/com/paichai/health/ptsession/entity/PtSession.java
@@ -1,0 +1,5 @@
+package com.paichai.health.ptsession.entity;
+
+public class PtSession {
+
+}
