+++ backend/src/main/java/com/paichai/health/ptsession/repository/PtSessionRepository.java
@@ -1,0 +1,5 @@
+package com.paichai.health.ptsession.repository;
+
+public class PtSessionRepository {
+
+}
