+++ backend/src/main/java/com/paichai/health/ptsession/service/PtSessionService.java
@@ -1,0 +1,5 @@
+package com.paichai.health.ptsession.service;
+
+public class PtSessionService {
+
+}
