--- backend/src/main/java/com/paichai/health/routine_review/controller/RoutineReviewController.java
@@ -1,5 +1,0 @@
-package com.paichai.health.routine_review.controller;
-
-public class RoutineReviewController {
-
-}
