--- backend/src/main/java/com/paichai/health/routine_review/dto/RoutineReviewRequest.java
@@ -1,5 +1,0 @@
-package com.paichai.health.routine_review.dto;
-
-public class RoutineReviewRequest {
-
-}
