--- backend/src/main/java/com/paichai/health/routine_review/dto/RoutineReviewResponse.java
@@ -1,5 +1,0 @@
-package com.paichai.health.routine_review.dto;
-
-public class RoutineReviewResponse {
-
-}
