--- backend/src/main/java/com/paichai/health/routine_review/entity/RoutineReview.java
@@ -1,5 +1,0 @@
-package com.paichai.health.routine_review.entity;
-
-public class RoutineReview {
-
-}
