--- backend/src/main/java/com/paichai/health/routine_review/repository/RoutineReviewRepository.java
@@ -1,5 +1,0 @@
-package com.paichai.health.routine_review.repository;
-
-public interface RoutineReviewRepository {
-
-}
