--- backend/src/main/java/com/paichai/health/routine_review/service/RoutineReviewController.java
@@ -1,5 +1,0 @@
-package com.paichai.health.routine_review.service;
-
-public class RoutineReviewController {
-
-}
