+++ backend/src/main/java/com/paichai/health/routinereview/controller/RoutineReviewController.java
@@ -1,0 +1,5 @@
+package com.paichai.health.routinereview.controller;
+
+public class RoutineReviewController {
+
+}
