+++ backend/src/main/java/com/paichai/health/routinereview/dto/RoutineReviewRequest.java
@@ -1,0 +1,5 @@
+package com.paichai.health.routinereview.dto;
+
+public class RoutineReviewRequest {
+
+}
