+++ backend/src/main/java/com/paichai/health/routinereview/entity/RoutineReview.java
@@ -1,0 +1,5 @@
+package com.paichai.health.routinereview.entity;
+
+public class RoutineReview {
+
+}
