+++ backend/src/main/java/com/paichai/health/routinereview/repository/RoutineReviewRepository.java
@@ -1,0 +1,5 @@
+package com.paichai.health.routinereview.repository;
+
+public interface RoutineReviewRepository {
+
+}
