+++ backend/src/main/java/com/paichai/health/trainerclientrequest/controller/TrainerClientRequestController.java
@@ -1,0 +1,5 @@
+package com.paichai.health.trainerclientrequest.controller;
+
+public class TrainerClientRequestController {
+
+}
