+++ backend/src/main/java/com/paichai/health/trainerclientrequest/dto/TrainerClientRequestRequest.java
@@ -1,0 +1,5 @@
+package com.paichai.health.trainerclientrequest.dto;
+
+public class TrainerClientRequestRequest {
+
+}
