+++ backend/src/main/java/com/paichai/health/trainerclientrequest/entity/TrainerClientRequest.java
@@ -1,0 +1,5 @@
+package com.paichai.health.trainerclientrequest.entity;
+
+public class TrainerClientRequest {
+
+}
