+++ backend/src/main/java/com/paichai/health/trainerclientrequest/repository/TrainerClientRequestRepository.java
@@ -1,0 +1,5 @@
+package com.paichai.health.trainerclientrequest.repository;
+
+public class TrainerClientRequestRepository {
+
+}
