+++ backend/src/main/java/com/paichai/health/trainerclientrequest/service/TrainerClientRequestService.java
@@ -1,0 +1,5 @@
+package com.paichai.health.trainerclientrequest.service;
+
+public class TrainerClientRequestService {
+
+}
