+++ backend/src/main/java/com/paichai/health/userroutine/controller/UserRoutineController.java
@@ -1,0 +1,5 @@
+package com.paichai.health.userroutine.controller;
+
+public class UserRoutineController {
+
+}
