+++ backend/src/main/java/com/paichai/health/userroutine/dto/UserRoutineRequest.java
@@ -1,0 +1,5 @@
+package com.paichai.health.userroutine.dto;
+
+public class UserRoutineRequest {
+
+}
