+++ backend/src/main/java/com/paichai/health/userroutine/entity/UserRoutine.java
@@ -1,0 +1,5 @@
+package com.paichai.health.userroutine.entity;
+
+public class UserRoutine {
+
+}
