+++ backend/src/main/java/com/paichai/health/userroutine/repository/UserRoutineRepository.java
@@ -1,0 +1,5 @@
+package com.paichai.health.userroutine.repository;
+
+public class UserRoutineRepository {
+
+}
