+++ backend/src/main/java/com/paichai/health/userroutine/service/UserRoutineService.java
@@ -1,0 +1,5 @@
+package com.paichai.health.userroutine.service;
+
+public class UserRoutineService {
+
+}
